git-svn-id: https://www.seasar.org/svn/s2dao.php5/trunk@214 18cb923b-d800-0410-ad6b-a912cf186b9e

+++ src/s2dao.php5/org/seasar/dao/activerecord/S2ActiveRecord.class.php
@@ -9,6 +9,10 @@ abstract class S2ActiveRecord {
     protected $row = array();
     private $map = null;
     
+    const regex_insert_names = '/^(insert|add|create)/';
+    const regex_update_names = '/^(update|modify|store)/';
+    const regex_delete_names = '/^(delete|remove)/';
+    
     public function __construct(S2Container_DataSource $dataSource){
         $this->helper = new S2ActiveRecordHelper($dataSource,
                                                 new ReflectionClass(get_class($this)));
@@ -48,6 +52,16 @@ abstract class S2ActiveRecord {
         $this->row[$column] = $value;
     }
     
+    public function __toString(){
+        $str = '[TABLE]: ' . $this->helper->getTable() . PHP_EOL;
+        $str .= '[COLUMNS]: ';
+        $cols = array();
+        foreach($this->row as $column => $value){
+            $cols[] = $column . ' = ' . $value;
+        }
+        return $str . implode(',', $cols);
+    }
+    
     public function __call($name, $param){
         $type = '';
         $attr = '';
@@ -59,11 +73,19 @@ abstract class S2ActiveRecord {
             $this->__set($attr, $param[0]);
         } else if($type == 'get'){
             return $this->__get($attr);
-        } else if($this->helper->isRecursiveMethod($name)){
+        } else if(preg_match(self::regex_insert_names, $name)){
+            return $this->_insert();
+        } else if(preg_match(self::regex_update_names, $name)){
+            return $this->_update();
+        } else if(preg_match(self::regex_delete_names, $name)){
+            return $this->_delete();
         } else {
             $sql = $this->helper->getMethodSql($name);
             if($sql == null){
                 throw new Exception();
+            }
+            
+            if($this->helper->isRecursiveMethod($name)){
             }
         }
     }
@@ -72,30 +94,13 @@ abstract class S2ActiveRecord {
         $this->row = array();
     }
     
-    public function __toString(){
-        $str = 'TABLE: ' . $this->helper->getTable() . PHP_EOL;
-        $str .= 'COLUMNS: ' . PHP_EOL;
-        foreach($this->row as $column => $value){
-            $str .= $column . " = " . $value . PHP_EOL;
-        }
-        return $str;
-    }
-    
     public function toString(){
         return $this->__toString();
     }
     
-    protected function getPlaceholders(array $row){
-        $folder = array();
-        foreach($row as $column => $value){
-            $folder[] = $this->helper->getTable() . '.' .
-                        $this->map->get($column) . " = ?"; 
-        }
-        return $folder;
-    }
-    
     public function find(){
-        $sql = $this->getNormalSelectSql();
+        $args = func_get_args();
+        $sql = call_user_func_array(array(__CLASS__, 'createSelectSql'), $args);
         $row = $this->execute($sql)->fetch(PDO::FETCH_ASSOC);
         $row = array_change_key_case($row, CASE_LOWER);
         $cloneable = clone $this;
@@ -104,7 +109,8 @@ abstract class S2ActiveRecord {
     }
     
     public function findAll(){
-        $sql = $this->getNormalSelectSql();
+        $args = func_get_args();
+        $sql = call_user_func_array(array(__CLASS__, 'createSelectSql'), $args);
         $stmt = $this->execute($sql);
         $rows = $stmt->fetchAll(PDO::FETCH_ASSOC);
         $retVal = array();
@@ -125,11 +131,68 @@ abstract class S2ActiveRecord {
         }
     }
     
-    public function delete(){
-        if(!$this->isSetPkey()){
-            throw new Exception("Primery Key not set");
+    public function count(){
+        $sql = 'SELECT count(*) FROM ' . $this->helper->getTable();
+        return (int)$this->helper->query($sql)->fetchColumn();
+    }
+
+    protected function getPlaceholders(array $row){
+        $folder = array();
+        foreach($row as $column => $value){
+            $folder[] = $this->helper->getTable() . '.' .
+                        $this->map->get($column) . ' = ?'; 
         }
-        return $this->_delete();
+        return $folder;
+    }
+    
+    protected function createSelectSql(){
+        switch(func_num_args()){
+            case 0:
+                $sql = $this->createSelectSqlByNormal();
+                return $this->createSelectSql($sql, null);
+            case 1:
+                return $this->createSelectSql(func_get_arg(0), null);
+            case 2:
+                $args = func_get_args();
+                if($args[0] instanceof S2ActiveRecord){
+                    $sql = $this->createSelectSqlByCondition($args[0]);
+                    return $this->createSelectSql($sql, $args[1]);
+                } else if(!isset($args[0])){
+                    $sql = $this->createSelectSqlByNormal();
+                    return $this->createSelectSql($sql, $args[1]);
+                } else {
+                    $sql = $args[0];
+                    $order = $args[1];
+                    if(isset($order)){
+                        $sql .= $this->createOrderByNormal($order);
+                    }
+                    return $sql;
+                }
+            default:
+                return null;
+        }
+    }
+
+    protected function createSelectSqlByNormal(){
+        $sql = 'SELECT ' . implode(',', $this->helper->getColumnNames()) .
+               ' FROM ' . $this->helper->getTable();
+        if(0 < count($this->row)){
+            $sql .= ' WHERE ';
+            $sql .= implode(',', $this->getPlaceholders($this->row));
+        }
+        return $sql;
+    }
+    
+    protected function createSelectSqlByCondition(S2ActiveRecord $record){
+    }        
+
+    protected function createOrderByNormal(array $orders){
+        $order = array();
+        foreach($orders as $column => $sort){
+            $s = $this->helper->getTable() . '.' . $this->map->get($column);
+            $order[] = $s . ' ' . $sort;
+        }
+        return ' ORDER BY ' . implode(',', $order);
     }
     
     protected function isSetPkey(){
@@ -148,16 +211,6 @@ abstract class S2ActiveRecord {
         $this->helper->bindArgs($stmt, $this->row);
         $stmt->execute();
         return $stmt;
-    }
-    
-    protected function getNormalSelectSql(){
-        $sql = 'SELECT ' . implode(',', $this->helper->getColumnNames()) .
-               ' FROM ' . $this->helper->getTable();
-        if(0 < count($this->row)){
-            $sql .= ' WHERE ';
-            $sql .= implode(',', $this->getPlaceholders($this->row));
-        }
-        return $sql;
     }
     
     private function _insert(){
@@ -202,6 +255,10 @@ abstract class S2ActiveRecord {
     
     
     private function _delete(){
+        if(!$this->isSetPkey()){
+            throw new Exception("Primery Key not set");
+        }
+        
         $sql = 'DELETE FROM ' . $this->helper->getTable();
         $pkeys = $this->helper->getPrimaryKeyNames();
         $where = array();
@@ -224,6 +281,7 @@ abstract class S2ActiveRecord {
         $stmt->execute();
         return $stmt->rowCount();
     }
+    
 }
 
 ?>
